+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Rely on Proton to declare common annotation processor options.
+
 ### [v0.226](https://github.com/react4j/react4j/tree/v0.226) (2026-06-16) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.225...v0.226)
 
 Changes in this release: